Add orders and the ability to change a cart to an order
herefor we created a Order model with a migration, the order has many
line items and line items belong to an order. For the event of ordering
we created a Registration module in the model folder, that has a
function that changes the cart to an order with all the products. this
is done in a transaction in order to not go further after one mistake
and cancels the function.